api/security.py
# api/security.py

from datetime import datetime, timedelta, timezone
from typing import Optional
from jose import JWTError, jwt
from passlib.context import CryptContext

# This remains the same
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# --- NEW: JWT Configuration ---
# This is a secret key used to sign the JWTs. It should be a long, random
# string and must be kept secret.
# For production, this should come from your environment secrets.
# You can generate a good key with the command: openssl rand -hex 32
SECRET_KEY = "durins-day" # Replace this!
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30


def verify_password(plain_password: str, hashed_password: str) -> bool:
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password: str) -> str:
    return pwd_context.hash(password)

# --- NEW: JWT Functions ---
def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    """Creates a new JWT access token."""
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now(timezone.utc) + expires_delta
    else:
        expire = datetime.now(timezone.utc) + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt